refactor(ModalForm): rename providers/ModalForm to ModalFormProvider and decouple its context to a new file
