+++ src/context/ModalFormContext/ModalFormContext.js
@@ -1,0 +1,9 @@
+import { createContext } from "react";
+
+export const ModalFormContext = createContext({
+  handleOnClickCreateMeeting: () => {},
+  handleOnClickEditMeeting: () => {},
+  handleOnCloseModal: () => {},
+  handleDeleteMeeting: () => {},
+  handleCloseDialog: () => {}
+});
