+++ src/context/ModalFormContext/index.js
@@ -1,0 +1,1 @@
+export { ModalFormContext } from "./ModalFormContext";
