--- src/providers/ModalForm/index.js
@@ -1,1 +1,0 @@
-export { ModalFormConsumer, ModalFormProvider } from "./ModalForm";
