+++ src/providers/ModalFormProvider/index.js
@@ -1,0 +1,1 @@
+export { ModalFormConsumer, ModalFormProvider } from "./ModalFormProvider";
